pipeline/requirements.txt
sqlalchemy==2.0.38
requests==2.32.4
tika==2.6.0
meilisearch==0.31.0
psycopg2-binary==2.9.10
google-genai==0.1.0
prometheus-client==0.19.0
spacy==3.7.4
camelot-py==0.11.0
opencv-python-headless==4.9.0.80
pypdf==6.6.2
scikit-learn==1.5.0
rapidfuzz==3.12.1
pytest==8.0.0
pytest-mock==3.12.0
httpx==0.28.1
